+++ R/find_data.R
@@ -3,8 +3,8 @@
 #' @param model The model object.
 #' @param \dots Additional arguments passed to methods.
 #' @param env An environment in which to look for the \code{data} argument to the modelling call.
-#' @details This typically requires that a model object be specified using a formula interface and an explicit \code{data} argument. Models that can be specified using variables from the \code{.GlobalEnv} or with a non-formula interface (e.g., a matrix of data) will tend to generate errors.
-#' @return A data frame, typically with one column unless the variable is a factor with more than two levels.
+#' @details This is a convenience function and, as such, carries no guarantees. To behave well, it typically requires that a model object be specified using a formula interface and an explicit \code{data} argument. Models that can be specified using variables from the \code{.GlobalEnv} or with a non-formula interface (e.g., a matrix of data) will tend to generate errors. \code{find_data} is an S3 generic so it is possible to expand it with new methods.
+#' @return A data frame containing the original data used in a modelling call, modified according to the original model's `subset` and `na.action` arguments, if appropriate.
 #' @examples
 #' require("datasets")
 #' x <- lm(mpg ~ cyl * hp + wt, data = head(mtcars))
@@ -36,6 +36,9 @@ find_data.default <- function(model, env = parent.frame(), ...) {
         if (!is.null(model[["na.action"]])) {
             dat <- dat[-model[["na.action"]], , drop = FALSE]
         }
+    }
+    if (is.null(dat)) {
+        stop("'find_data()' requires a formula call")
     }
     dat
 }
@@ -82,7 +85,7 @@ find_data.vgam <- function(model, env = parent.frame(), ...) {
     if (!requireNamespace("methods")) {
         stop("'find_data.vgam()' requires the 'methods' package")
     }
-    dat <- methods::slot(hfit, "misc")[["dataname"]]
+    dat <- methods::slot(model, "misc")[["dataname"]]
     get(dat, envir = env)
 }