+++ R/prediction_clm.R
@@ -62,7 +62,7 @@ function(model,
               row.names = seq_len(nrow(pred)),
               at = if (is.null(at)) at else names(at), 
               model.class = class(model),
-              type = NA_character,
+              type = NA_character_,
               category = category)
 }